refactor: infrastructure 패키지 이동
