+++ backend/src/main/java/harustudy/backend/auth/domain/oauth/OauthClient.java
@@ -1,4 +1,4 @@
-package harustudy.backend.auth.infrastructure;
+package harustudy.backend.auth.domain.oauth;
 
 import harustudy.backend.auth.dto.OauthTokenResponse;
